pyproject.toml
[project]
name = "api_bancaria_assincrona"
version = "0.1.0"
description = ""
authors = [
    {name = "CristianoLeitao",email = "cristianoleitaolima@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"

[tool.poetry.dependencies]
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = "^2.0.30"
aiosqlite = "^0.20.0"
python-jose = "^3.3.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pydantic = "^2.7.0"
bcrypt = "4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"
black = "^24.4.0"
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
